Move segments permissions tests into unit test file
